rag_ingestion/requirements.txt
# Kindred RAG Ingestion Pipeline Dependencies

# Pinecone vector database client with integrated embedding
pinecone>=5.0.0

# Tiktoken for accurate token counting (chunking)
tiktoken>=0.5.0

# Load environment variables from .env files
python-dotenv>=1.0.0
